[FE] FIX: axios instance 쿠키 수정 #877
