--- frontend/src/api/axios/axios.admin.custom.ts
@@ -1,11 +1,0 @@
-import adminInstance from "./axios.admin.instance";
-
-const axiosLentURL = "/api/admin/lent";
-export const axiosLent = async () => {
-  try {
-    const response = await adminInstance.get(axiosLentURL);
-    return response;
-  } catch (error) {
-    throw error;
-  }
-};
