+++ frontend/src/api/axios/axios.custom.ts
@@ -119,12 +119,23 @@ export const axiosReturn = async (): Promise<any> => {
   }
 };
 
-const axiosMyLentLogURL = "api/my_lent_info/log";
+const axiosMyLentLogURL = "/api/my_lent_info/log";
 export const axiosMyLentLog = async (page: number): Promise<any> => {
   try {
     const response = await instance.get(axiosMyLentLogURL, {
       params: { page: page, length: 10 },
     });
+    return response;
+  } catch (error) {
+    throw error;
+  }
+};
+
+// Admin API
+const axiosLentURL = "/api/admin/lent";
+export const axiosLent = async () => {
+  try {
+    const response = await instance.get(axiosLentURL);
     return response;
   } catch (error) {
     throw error;
